[#176] 맵 기믹 컴포넌트 분리, 낙사 가능한 벽 소스코드 생성
