+++ Source/Project_LLL/Private/System/MapGimmick/Components/LLL_GateSpawnPointComponent.cpp
@@ -1,0 +1,36 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+
+#include "System/MapGimmick/Components/LLL_GateSpawnPointComponent.h"
+#include "Components/SphereComponent.h"
+
+// Sets default values for this component's properties
+ULLL_GateSpawnPointComponent::ULLL_GateSpawnPointComponent()
+{
+	// Set this component to be initialized when the game starts, and to be ticked every frame.  You can turn these features
+	// off to improve performance if you don't need them.
+	PrimaryComponentTick.bCanEverTick = false;
+
+	// ...
+	CollisionSphere = CreateDefaultSubobject<USphereComponent>(TEXT("VisibleCollision"));
+}
+
+
+// Called when the game starts
+void ULLL_GateSpawnPointComponent::BeginPlay()
+{
+	Super::BeginPlay();
+
+	// ...
+	
+}
+
+
+// Called every frame
+void ULLL_GateSpawnPointComponent::TickComponent(float DeltaTime, ELevelTick TickType, FActorComponentTickFunction* ThisTickFunction)
+{
+	Super::TickComponent(DeltaTime, TickType, ThisTickFunction);
+
+	// ...
+}
+
