+++ Source/Project_LLL/Private/System/MapGimmick/Components/LLL_ProductSpawnPointComponent.cpp
@@ -1,0 +1,34 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+
+#include "System/MapGimmick/Components/LLL_ProductSpawnPointComponent.h"
+
+// Sets default values for this component's properties
+ULLL_ProductSpawnPointComponent::ULLL_ProductSpawnPointComponent()
+{
+	// Set this component to be initialized when the game starts, and to be ticked every frame.  You can turn these features
+	// off to improve performance if you don't need them.
+	PrimaryComponentTick.bCanEverTick = false;
+
+	// ...
+}
+
+
+// Called when the game starts
+void ULLL_ProductSpawnPointComponent::BeginPlay()
+{
+	Super::BeginPlay();
+
+	// ...
+	
+}
+
+
+// Called every frame
+void ULLL_ProductSpawnPointComponent::TickComponent(float DeltaTime, ELevelTick TickType, FActorComponentTickFunction* ThisTickFunction)
+{
+	Super::TickComponent(DeltaTime, TickType, ThisTickFunction);
+
+	// ...
+}
+
