+++ Source/Project_LLL/Private/System/MapGimmick/LLL_MapGimmick.cpp
@@ -12,9 +12,9 @@
 #include "Entity/Character/Player/LLL_PlayerBase.h"
 #include "Entity/Object/Interactive/Gate/LLL_GateObject.h"
 #include "Entity/Object/Interactive/Reward/LLL_RewardObject.h"
-#include "System/MapGimmick/LLL_GateSpawnPointComponent.h"
-#include "System/MapGimmick/LLL_ShoppingMapComponent.h"
-#include "System/MapGimmick/LLL_PlayerSpawnPointComponent.h"
+#include "System/MapGimmick/Components/LLL_GateSpawnPointComponent.h"
+#include "System/MapGimmick/Components/LLL_ShoppingMapComponent.h"
+#include "System/MapGimmick/Components/LLL_PlayerSpawnPointComponent.h"
 #include "System/MonsterSpawner/LLL_MonsterSpawner.h"
 #include "System/Reward/LLL_RewardGimmick.h"
 #include "Util/LLL_ConstructorHelper.h"
