+++ Source/Project_LLL/Public/System/MapGimmick/Components/LLL_GateSpawnPointComponent.h
@@ -1,0 +1,31 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+#pragma once
+
+#include "CoreMinimal.h"
+#include "Components/SceneComponent.h"
+#include "LLL_GateSpawnPointComponent.generated.h"
+
+class USphereComponent;
+
+UCLASS( ClassGroup=(Custom), meta=(BlueprintSpawnableComponent) )
+class PROJECT_LLL_API ULLL_GateSpawnPointComponent : public USceneComponent
+{
+	GENERATED_BODY()
+
+public:	
+	// Sets default values for this component's properties
+	ULLL_GateSpawnPointComponent();
+
+protected:
+	// Called when the game starts
+	virtual void BeginPlay() override;
+
+public:	
+	// Called every frame
+	virtual void TickComponent(float DeltaTime, ELevelTick TickType, FActorComponentTickFunction* ThisTickFunction) override;
+
+protected:
+	UPROPERTY(EditDefaultsOnly)
+	TObjectPtr<USphereComponent> CollisionSphere;
+};
