+++ Source/Project_LLL/Public/System/MapGimmick/LLL_FallableWallGimmick.h
@@ -1,0 +1,27 @@
+﻿// Fill out your copyright notice in the Description page of Project Settings.
+
+#pragma once
+
+#include "CoreMinimal.h"
+#include "System/Base/LLL_SystemBase.h"
+#include "LLL_FallableWallGimmick.generated.h"
+
+UCLASS()
+class PROJECT_LLL_API ALLL_FallableWallGimmick : public ALLL_SystemBase
+{
+	GENERATED_BODY()
+
+public:
+	// Sets default values for this actor's properties
+	ALLL_FallableWallGimmick();
+
+protected:
+	// Called when the game starts or when spawned
+	virtual void BeginPlay() override;
+	virtual void NotifyHit(UPrimitiveComponent* MyComp, AActor* Other, UPrimitiveComponent* OtherComp, bool bSelfMoved, FVector HitLocation, FVector HitNormal, FVector NormalImpulse, const FHitResult& Hit) override;
+	
+protected:
+	UPROPERTY()
+	TObjectPtr<UStaticMeshComponent> Wall;
+	
+};
